PACMAN_CACHE environment variable sets an optional local package cache.
This variable may be set to the location of a cache directory outside the build
chroot. If set, all files in the cache directory are copied to the chroot
/var/cache/pacman/pkg before all chroot pacman operations. After chroot pacman
operations are finished, all files in chroot /var/cache/pacman/pkg are copied
to the cache directory. This avoids repeatedly syncing packages across builds,
which can improve build performance.